Allow replacing a task.
ArtifactHolder.replace indicates that the task is now the one
generating the artifact. This means the default task that
normally generates it does not need to be created.

Right now it does not handle the input of the replacement task
unlike the transform task. Obviously this task is going to consume
something, so we need to handle this somehow.