move single transform to a new direction.
The API now does not take over registering the task,
and does not require the transform task to extend anything particular.
Instead the API receives the TaskProvider as well as 2 lambdas to lazily
provide the input and output.

This offer greater flexibility when creating tasks that need to transform
more than one artifact.